requirements.txt
PySimpleGUI==4.20.0
python-jenkins==1.7.0